Move the logic of the RecyclerView adapter listener to ArticleRecyclerViewAdapter.kt file
